Project-2/project_2/requirements.txt
flask
<<<<<<< HEAD
psycopg2-binary
=======
requests
>>>>>>> 296ad6650222a786eb98c7a05bbf7e7ed03bd999

